TermoSmart/termosmart-func/requirements.txt
azure-functions
azure-storage-blob
azure-core
bcrypt
